organize merge file structure
