+++ crates/deltalake-core/src/delta_datafusion/mod.rs
@@ -81,7 +81,6 @@ use crate::{open_table, open_table_with_storage_options, DeltaTable};
 
 const PATH_COLUMN: &str = "__delta_rs_path";
 
-pub mod barrier;
 pub mod expr;
 pub mod logical;
 pub mod physical;
